+++ models/chat.go
@@ -65,7 +65,7 @@ type ChatLocation struct {
 
 // Chat https://core.telegram.org/bots/api#chat
 type Chat struct {
-	ID                                 int              `json:"id"`
+	ID                                 int64            `json:"id"`
 	Type                               string           `json:"type"`
 	Title                              string           `json:"title,omitempty"`
 	Username                           string           `json:"username,omitempty"`
@@ -86,6 +86,6 @@ type Chat struct {
 	HasProtectedContent                bool             `json:"has_protected_content,omitempty"`
 	StickerSetName                     string           `json:"sticker_set_name,omitempty"`
 	CanSetStickerSet                   bool             `json:"can_set_sticker_set,omitempty"`
-	LinkedChatID                       int              `json:"linked_chat_id,omitempty"`
+	LinkedChatID                       int64            `json:"linked_chat_id,omitempty"`
 	Location                           *ChatLocation    `json:"location,omitempty"`
 }
